Gui: Make Document independent of QWidget view
To use Gui::Document without the mainwindow it must not only work with View3DInventor, which is a QWidget, but also with widgetless viewers. Hence the scenegraph logic of the viewer is splitted from the widgets to allow for widgetless scenegraph viewers and Gui::Document is changed to work with this widgetless classes.